Checks for a loop within a linked list
Checks to see if there i a loop within a linked list by having two nodes. The first node is on the current node, the second is on node.next. By having these two pointers if there is a loop, both would eventually be equal to each other. If there is no loop, the second node would be null, which would indicate there is no loop.